requirements.txt
torch==2.8.0
torchvision==0.23.0
diffusers
transformers
peft
accelerate
bitsandbytes
safetensors
wandb
matplotlib
requests
pillow
tqdm
openai
PyYAML